feat: enhance the search section
