+++ src/components/search/SearchBar/AdvancedSearchBar.tsx
@@ -1,7 +1,14 @@
 'use client';
 
+import Pagination, {
+  PaginationItem,
+  PaginationNext,
+  PaginationPrev,
+  generatePaginationItems,
+} from '@components/Pagination';
+import { cva } from 'class-variance-authority';
 import { Map } from 'lucide-react';
-import { useRouter } from 'next/navigation';
+import { useRouter, useSearchParams } from 'next/navigation';
 import { useState } from 'react';
 import AdvancedSearchBarDesktop from './AdvancedSearchBarDesktop';
 import AdvancedSearchBarMobile from './AdvancedSearchBarMobile';
@@ -9,8 +16,28 @@ import { useSearchProvider } from './SearchProvider';
 import { useSearchView } from './SearchViewProvider';
 import { SearchParams, updateQueryString } from './fields/utils';
 
-const AdvancedSearchBar = () => {
+type AdvancedSearchBarProps = {
+  totalPage: number;
+  currentPage: number;
+  children?: React.ReactNode;
+};
+
+const paginationStepperStyles = cva('flex gap-4 py-12', {
+  variants: {
+    isMobile: {
+      true: 'justify-center px-[6.46875rem] gap-12',
+      false: 'justify-between px-[8.031rem]',
+    },
+  },
+});
+
+const AdvancedSearchBar = ({
+  totalPage,
+  currentPage,
+  children,
+}: AdvancedSearchBarProps) => {
   const router = useRouter();
+  const searchParams = useSearchParams();
   const [scrollPosition, setScrollPosition] = useState(0);
   const { isMobile, ...formMethods } = useSearchProvider<SearchParams>();
   const { currentShow, setCurrentShow } = useSearchView();
@@ -51,28 +78,84 @@ const AdvancedSearchBar = () => {
     setCurrentShow((prev) => (prev === 'results' ? 'map' : 'results'));
   };
 
-  return isMobile ? (
-    <section className="fixed bottom-0 left-0 right-0 z-10 flex justify-between gap-2 border-t border-primary bg-surface px-3 py-4">
-      <AdvancedSearchBarMobile
-        onSearchSubmit={handleSearchSubmit}
-        onClearFilter={() => handleClearFilter}
-      />
-      <button
-        type="button"
-        aria-label="Map button"
-        className="h-14 w-14 border-[1px] border-primary bg-surface p-4 font-medium text-primary"
-        onClick={() => {
-          toggleCurrentShow();
-        }}
+  const handlePageChange = (newPage: number) => {
+    const updatedSearchParams = new URLSearchParams(searchParams);
+    updatedSearchParams.set('page', newPage.toString());
+    router.push(`/search?${updatedSearchParams.toString()}`);
+  };
+
+  const renderPaginationItems = () => {
+    return generatePaginationItems(currentPage, totalPage).map((item) => {
+      if (item.type === PaginationItem) {
+        return (
+          <PaginationItem
+            key={item.key}
+            page={item.props.page}
+            isCurrent={item.props.isCurrent}
+            onClick={() => handlePageChange(item.props.page)}
+          />
+        );
+      }
+      return item;
+    });
+  };
+
+  const renderPagination = () => (
+    <div id="results-pagination" className="w-full xl:max-w-[53.5rem]">
+      <Pagination
+        currentPage={currentPage}
+        totalPage={totalPage}
+        onClickPrev={() => handlePageChange(Math.max(currentPage - 1, 1))}
+        onClickNext={() =>
+          handlePageChange(Math.min(currentPage + 1, totalPage))
+        }
+        asChild
       >
-        <Map />
-      </button>
-    </section>
-  ) : (
-    <AdvancedSearchBarDesktop
-      onSearchSubmit={handleSearchSubmit}
-      onClearFilter={() => handleClearFilter}
-    />
+        <div
+          id="pagination-stepper"
+          className={paginationStepperStyles({ isMobile })}
+        >
+          <PaginationPrev
+            className={isMobile ? 'border-[1px]' : ''}
+            iconClassName={isMobile ? 'size-12' : ''}
+          />
+          {isMobile ? null : renderPaginationItems()}
+          <PaginationNext
+            className={isMobile ? 'border-[1px]' : ''}
+            iconClassName={isMobile ? 'size-12' : ''}
+          />
+        </div>
+      </Pagination>
+    </div>
+  );
+
+  return (
+    <>
+      {isMobile ? (
+        <section className="fixed bottom-0 left-0 right-0 z-10 flex justify-between gap-2 border-t border-primary bg-surface px-3 py-4">
+          <AdvancedSearchBarMobile
+            onSearchSubmit={handleSearchSubmit}
+            onClearFilter={handleClearFilter}
+          />
+          <button
+            type="button"
+            aria-label="Map button"
+            className="h-14 w-14 border-[1px] border-primary bg-surface p-4 font-medium text-primary"
+            onClick={toggleCurrentShow}
+          >
+            <Map />
+          </button>
+        </section>
+      ) : (
+        <AdvancedSearchBarDesktop
+          onSearchSubmit={handleSearchSubmit}
+          onClearFilter={handleClearFilter}
+        />
+      )}
+
+      {children}
+      {renderPagination()}
+    </>
   );
 };
